+++ src/main/java/com/example/bookiibookii/domain/support/inquiry/controller/InquiryController.java
@@ -1,8 +1,8 @@
-package com.example.bookiibookii.domain.support.controller;
+package com.example.bookiibookii.domain.support.inquiry.controller;
 
-import com.example.bookiibookii.domain.support.dto.req.InquiryRequestDTO;
-import com.example.bookiibookii.domain.support.dto.res.InquiryResponseDTO;
-import com.example.bookiibookii.domain.support.service.InquiryService;
+import com.example.bookiibookii.domain.support.inquiry.dto.req.InquiryRequestDTO;
+import com.example.bookiibookii.domain.support.inquiry.dto.res.InquiryResponseDTO;
+import com.example.bookiibookii.domain.support.inquiry.service.InquiryService;
 import com.example.bookiibookii.domain.user.entity.User;
 import com.example.bookiibookii.global.apiPayload.ApiResponse;
 import com.example.bookiibookii.global.apiPayload.code.GeneralSuccessCode;
