+++ src/main/java/com/example/bookiibookii/domain/support/inquiry/controller/InquiryControllerDocs.java
@@ -1,7 +1,7 @@
-package com.example.bookiibookii.domain.support.controller;
+package com.example.bookiibookii.domain.support.inquiry.controller;
 
-import com.example.bookiibookii.domain.support.dto.req.InquiryRequestDTO;
-import com.example.bookiibookii.domain.support.dto.res.InquiryResponseDTO;
+import com.example.bookiibookii.domain.support.inquiry.dto.req.InquiryRequestDTO;
+import com.example.bookiibookii.domain.support.inquiry.dto.res.InquiryResponseDTO;
 import com.example.bookiibookii.domain.user.entity.User;
 import com.example.bookiibookii.global.apiPayload.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
