+++ src/main/java/com/example/bookiibookii/domain/support/inquiry/enums/SupportStatus.java
@@ -1,4 +1,4 @@
-package com.example.bookiibookii.domain.support.enums;
+package com.example.bookiibookii.domain.support.inquiry.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
