+++ src/main/java/com/example/bookiibookii/domain/support/inquiry/repository/InquiryRepository.java
@@ -1,6 +1,6 @@
-package com.example.bookiibookii.domain.support.repository;
+package com.example.bookiibookii.domain.support.inquiry.repository;
 
-import com.example.bookiibookii.domain.support.entity.Inquiry;
+import com.example.bookiibookii.domain.support.inquiry.entity.Inquiry;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
