+++ src/main/java/com/example/bookiibookii/domain/support/report/dto/res/ReportResponseDTO.java
@@ -1,7 +1,7 @@
-package com.example.bookiibookii.domain.support.dto.res;
+package com.example.bookiibookii.domain.support.report.dto.res;
 
-import com.example.bookiibookii.domain.support.enums.SupportStatus;
-import com.example.bookiibookii.domain.support.enums.ReportType;
+import com.example.bookiibookii.domain.support.inquiry.enums.SupportStatus;
+import com.example.bookiibookii.domain.support.report.enums.ReportType;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 import java.time.LocalDateTime;
