+++ src/main/java/com/example/bookiibookii/domain/support/report/service/ReportService.java
@@ -1,13 +1,13 @@
-package com.example.bookiibookii.domain.support.service;
+package com.example.bookiibookii.domain.support.report.service;
 
 import com.example.bookiibookii.domain.group.entity.Groups;
 import com.example.bookiibookii.domain.group.exception.GroupException;
 import com.example.bookiibookii.domain.group.exception.code.GroupErrorCode;
 import com.example.bookiibookii.domain.group.repository.GroupsRepository;
-import com.example.bookiibookii.domain.support.dto.req.ReportRequestDTO;
-import com.example.bookiibookii.domain.support.dto.res.ReportResponseDTO;
-import com.example.bookiibookii.domain.support.entity.Report;
-import com.example.bookiibookii.domain.support.repository.ReportRepository;
+import com.example.bookiibookii.domain.support.report.dto.req.ReportRequestDTO;
+import com.example.bookiibookii.domain.support.report.dto.res.ReportResponseDTO;
+import com.example.bookiibookii.domain.support.report.entity.Report;
+import com.example.bookiibookii.domain.support.report.repository.ReportRepository;
 import com.example.bookiibookii.domain.user.entity.User;
 import com.example.bookiibookii.domain.user.exception.UserException;
 import com.example.bookiibookii.domain.user.exception.code.UserErrorCode;
